rename 'winning tickets' to 'orchestrator payouts'
